Перенес аватар в модуль пользователя.
