+++ backend/src/app.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { AvatarsModule } from './avatars/avatars.module';
 import { Config, getConfig } from './config/config';
 import { DictionaryModule } from './dictionary/dictionary.module';
 import { MailModule } from './mail/mail.module';
@@ -16,7 +15,6 @@ const url = `${config.DATABASE_URL}`;
     MailModule,
     UsersModule,
     TasksModule,
-    AvatarsModule,
     MongooseModule.forRoot(url, {
       useNewUrlParser: true
     }),
