+++ backend/src/users/services/avatars/default-avatars.service.ts
